requirements.txt
Django==3.1.7
psycopg2-binary==2.8.6
django-crispy-forms==1.11.2
django-allauth==0.44.0
django-environ==0.4.5
dj-database-url==0.5.0
pillow==7.2.0
django-debug-toolbar==3.2.1
whitenoise==5.2.0
gunicorn==20.1.0 